+++ service/conversations/v1/user/api_op_delete.go
@@ -1,0 +1,33 @@
+// Package user contains auto-generated files. DO NOT MODIFY
+package user
+
+import (
+	"context"
+	"net/http"
+
+	"github.com/RJPearson94/twilio-sdk-go/client"
+)
+
+// Delete removes a user resource from the account
+// See https://www.twilio.com/docs/conversations/api/user-resource#delete-an-user-resource for more details
+// Context is defaulted to Background. See https://golang.org/pkg/context/#Background for more information
+func (c Client) Delete() error {
+	return c.DeleteWithContext(context.Background())
+}
+
+// DeleteWithContext removes a user resource from the account
+// See https://www.twilio.com/docs/conversations/api/user-resource#delete-an-user-resource for more details
+func (c Client) DeleteWithContext(context context.Context) error {
+	op := client.Operation{
+		Method: http.MethodDelete,
+		URI:    "/Users/{sid}",
+		PathParams: map[string]string{
+			"sid": c.sid,
+		},
+	}
+
+	if err := c.client.Send(context, op, nil, nil); err != nil {
+		return err
+	}
+	return nil
+}