+++ CHANGELOG.md
@@ -16,6 +16,7 @@ FEATURES
 
 - Add Delivery response field and struct to message resources
 - **New Resource:** Roles
+- **New Resource:** Users
 
 ### Serverless